Partly replaced the timer class.
The sketch is now functional, though some tinkering is needed in order
to restore the individual games previous progression. As of this moment,
the asteroid game does not gradually add new asteroids and the tank game
does not increase the number of enemies, as both are tied to timer
functions that have not been fixed yet. Small steps.